desktop/requirements.txt
Pillow==9.5.0
openai==1.69.0
keyboard==0.13.5
pyperclip==1.8.2
python-dotenv==0.17.1
requests==2.25.1
flask==3.1.0
flask-sqlalchemy==3.1.1
flask-login==0.6.3
flask-bcrypt==1.0.1
stripe==2.60.0
pyinstaller==6.12.0
